Добавил минификацию js
