+++ gulpfile.js
@@ -11,9 +11,11 @@ var rename = require("gulp-rename");
 var imagemin = require("gulp-imagemin");
 var webp = require("gulp-webp");
 var svgstore = require("gulp-svgstore");
+var cheerio = require("gulp-cheerio")
 var posthtml = require("gulp-posthtml");
 var include = require("posthtml-include");
-var del = require("del")
+var del = require("del");
+var uglify = require("gulp-uglify");
 
 gulp.task("css", function () {
   return gulp.src("source/less/style.less")
@@ -36,6 +38,13 @@ gulp.task("html", function() {
     ]))
   .pipe(gulp.dest("build"))
 });
+
+gulp.task("uglify", function() {
+  return gulp.src("source/js/*.js")
+  .pipe(uglify())
+  .pipe(rename({suffix: '.min'}))
+  .pipe(gulp.dest("build/js"))
+})
 
 gulp.task("server", function () {
   server.init({
@@ -77,7 +86,7 @@ gulp.task("sprite", function() {
     .pipe(svgstore({
       inlineSvg: true
     }))
-    .pipe(rename("sprites.svg"))
+    .pipe(rename("sprite.svg"))
     .pipe(gulp.dest("build/img"))
 });
 
@@ -85,7 +94,7 @@ gulp.task("copy", function() {
   return gulp.src([
     "source/fonts/**/*.{woff,woff2}",
     "source/img/**",
-    "source/js/**"
+    "source/js/*.js"
     ], {
       base: "source"
     })
@@ -99,6 +108,7 @@ gulp.task("clean", function() {
 gulp.task("build", gulp.series(
   "clean",
   "copy",
+  "uglify",
   "css",
   "sprite",
   "html"
