+++ package.json
@@ -9,6 +9,7 @@
     "browser-sync": "2.24.x",
     "del": "^3.0.0",
     "gulp": "4.0.x",
+    "gulp-cheerio": "^0.6.3",
     "gulp-csso": "^3.0.1",
     "gulp-imagemin": "^4.1.0",
     "gulp-less": "4.0.x",
@@ -17,6 +18,8 @@
     "gulp-posthtml": "^3.0.4",
     "gulp-rename": "^1.4.0",
     "gulp-svgstore": "^7.0.0",
+    "gulp-uglify": "^3.0.1",
+    "gulp-uglify-js": "0.0.4",
     "gulp-webp": "^3.0.0",
     "posthtml-include": "^1.2.0",
     "stylelint": "9.6.x",
