+++ source/catalog.html
@@ -3,12 +3,12 @@
   <head>
     <meta charset="UTF-8">
     <title>Каталог товаров</title>
-    <link rel="stylesheet" href="css/style.css">
+    <link rel="stylesheet" href="css/style.min.css">
     <meta name="viewport" content="width=device-width">
   </head>
   <body>
     <div style="display: none;">
-      <include src="build/img/sprites.svg"></include>
+      <include src="build/img/sprite.svg"></include>
     </div>
     <div class="wrapper">
       <!--Шапка страницы-->
@@ -239,10 +239,9 @@
         <button class="modal__button button">Добавить</button>
       </div>
     </div>
-    <script src="js/polyfill.js"></script>
-    <script src="js/svg4everybody.js"></script>
+    <script src="js/polyfill.min.js"></script>
+    <script src="js/svg4everybody.min.js"></script>
     <script>svg4everybody();</script>
-    <script src="js/mobile-menu.js"></script>
-    <script src="js/popup.js"></script>
+    <script src="js/index.min.js"></script>
 </body>
 </html>
