+++ source/form.html
@@ -3,12 +3,12 @@
 <head>
   <meta charset="UTF-8">
   <title>Форма магазина</title>
-  <link rel="stylesheet" href="css/style.css">
+  <link rel="stylesheet" href="css/style.min.css">
   <meta name="viewport" content="width=device-width">
 </head>
 <body>
   <div style="display: none;">
-    <include src="build/img/sprites.svg"></include>
+    <include src="build/img/sprite.svg"></include>
   </div>
   <div class="wrapper">
     <!--Шапка страницы-->
@@ -231,9 +231,9 @@
       </div>
     </footer>
   </div>
-  <script src="js/polyfill.js"></script>
-  <script src="js/svg4everybody.js"></script>
+  <script src="js/polyfill.min.js"></script>
+  <script src="js/svg4everybody.min.js"></script>
   <script>svg4everybody();</script>
-  <script src="js/mobile-menu.js"></script>
+  <script src="js/index.min.js"></script>
 </body>
 </html>
