--- source/js/mobile-menu.js
@@ -1,15 +1,0 @@
-var nav = document.querySelector('.nav');
-var navToggle = document.querySelector('.nav__toggle');
-
-nav.classList.remove('nav--nojs');
-
-navToggle.addEventListener('click', function() {
-  console.log("hello");
-  if (nav.classList.contains('nav--closed')) {
-    nav.classList.remove('nav--closed');
-    nav.classList.add('nav--opened');
-  } else {
-    nav.classList.add('nav--closed');
-    nav.classList.remove('nav--opened');
-  }
-});
